Remove type.svelte file
